requirements.txt
aiogram==3.6.0
gspread==6.0.2
python-dotenv==1.0.1